variables

+++ scalar/src/lib.rs
@@ -2,7 +2,7 @@ pub fn sum(a: u8, b: u8) -> u8 {
     a + b
 }
 
-pub fn diff(a: i8, b: i8) -> i8 {
+pub fn diff(a: i16, b: i16) -> i16 {
     a - b
 }
 
@@ -10,10 +10,10 @@ pub fn pro(a: i8, b: i8) -> i8 {
     a * b
 }
 
-pub fn quo(a: i8, b: i8) -> i8 {
+pub fn quo(a: f32, b: f32) -> f32 {
     a / b
 }
 
-pub fn rem(a: i8, b: i8) -> i8 {
+pub fn rem(a: f32, b: f32) -> f32 {
     a % b
 }
